-n i made a fix to pods.cmake in libbot/bot2-vis, and am propagating it to drake.  (it really should get propagated everywhere).  the symptom was “no rule to make ___blah__ library” errors, which I have seen a lot.  Here is the commit log from my libbot commit:
resolved strange linking error for X11 libraries on ubuntu.  turned out to be a sort of core logic bug in pods.cmake (target_link_libraries wants a semi-colon delimited list of arguments, but pkg-config --libs returns a white-space delimited list).  ultimately, this was a long, hard battle to make it compile well on mac with homebrew. :)

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/trunk@7719 c9849af7-e679-4ec6-a44e-fc146a885bd3